🔧 Fix Kotlin compilation errors in NotificationService
🐛 Fixed Compilation Errors:
• Remove problematic SimpleNotificationService with missing functions
• Revert to NotificationService with complete HTTP polling implementation
• All missing functions (createNotificationChannels, createForegroundNotification, etc.) added
• Fixed all unresolved references and if-else expression issues

📱 HTTP Polling Implementation:
• Complete HTTP-based device registration
• 3-second polling for notifications
• Proper notification delivery confirmation
• Enhanced logging for debugging

✅ Android project should now compile successfully
🚀 Ready for APK build and testing
