+++ android/app/src/main/java/com/bildirimtelefon/app/receiver/NotificationReceiver.kt
@@ -4,7 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import com.bildirimtelefon.app.service.SimpleNotificationService
+import com.bildirimtelefon.app.service.NotificationService
 
 class NotificationReceiver : BroadcastReceiver() {
     
@@ -22,10 +22,10 @@ class NotificationReceiver : BroadcastReceiver() {
                     
                     Log.d(TAG, "Test bildirimi alındı: $title - $message")
                     
-                    // SimpleNotificationService'ten test notification göster
-                    if (SimpleNotificationService.isRunning) {
+                    // NotificationService'ten test notification göster
+                    if (NotificationService.isRunning) {
                         // Service aktifse, service üzerinden göster
-                        val serviceIntent = Intent(context, SimpleNotificationService::class.java)
+                        val serviceIntent = Intent(context, NotificationService::class.java)
                         serviceIntent.putExtra("action", "show_test_notification")
                         serviceIntent.putExtra("title", title)
                         serviceIntent.putExtra("message", message)
